fix(popup): clean deferred contexts and failure logs
